added project board, updated and delete funtionality

+++ ppmtool-react-client/src/actions/projectActions.js
@@ -36,8 +36,13 @@ export const getProject = (id, history) => async dispatch => {
     history.push("/dashboard");
   }
 };
+
 export const deleteProject = id => async dispatch => {
-  if (window.confirm("Are you sure? This will delete all data")) {
+  if (
+    window.confirm(
+      "Are you sure? This will delete the project and all the data related to it"
+    )
+  ) {
     await axios.delete(`/api/project/${id}`);
     dispatch({
       type: DELETE_PROJECT,
